updated demos/local-integrated-setup scripts

+++ demos/local-integrated-setup/kill_all.sh
@@ -1,12 +1,31 @@
 #!/bin/sh 
 
 
-## kill osmosis process if running
-pkill osmosis 
+qpid=`ps -ef | grep quasarnoded | grep -v "grep" |  awk '{ printf $2 }'`
+if [ -z "$qpid" ]
+then
+      echo "quasarnoded not running."
+else
+      echo "quasarnoded is running with process id $qpid - killing" 
+ 	  pkill quasarnoded  
+fi
 
-## kill cosmos gaid process if running
-pkill gaiad
+opid=`ps -ef | grep osmosisd | grep -v "grep" |  awk '{ printf $2 }'`
+if [ -z "$opid" ]
+then
+      echo "osmosisd not running"
+else
+      echo "osmosisd is running with process id $opid - killing"
+	  pkill osmosisd
+fi
 
-## kill quasarnoded process if running
-pkill quasarnoded
+gpid=`ps -ef | grep "gaiad" | grep -v "grep" |  awk '{ printf $2 }'`
+if [ -z "$gpid" ]
+then
+      echo "gaiad not running"
+else
+      echo "gaiad is running with process id $gpid - killing"
+	  pkill gaiad
+fi
+
